requirements.txt
streamlit==1.38.0
plotly==5.24.0
pandas==2.2.2
numpy==1.26.4
folium==0.17.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
haversine==2.7.0
streamlit-folium
Pillow==10.3.0